requirements.txt
langchain==0.2.16
langchain-community==0.2.16
langchain-groq==0.1.9
chromadb==0.5.3
sentence-transformers==3.0.1
streamlit==1.37.0
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4